Update Eva+ Add ChatBox

+++ OpusLink.Entity/Models/FeedbackUser.cs
@@ -12,7 +12,7 @@ namespace OpusLink.Entity.Models
         public int JobID { get; set; }
         public int CreateByUserID { get; set; }
         public int TargetToUserID { get; set; }
-        public int Star { get; set; }
+        public decimal Star { get; set; }
         public string Content { get; set; }
         public DateTime DateCreated { get; set; }
